VOICE_GENERATOR/requirements.txt
flask==3.0.3
numpy==1.26.4
coqui-tts==0.27.5
